miner_sdk/pyproject.toml
[project]
name = "miner_sdk"
version = "0.1.0"
description = "A Python SDK for the ML-Chain Notary Network"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
    { name = "ML-Chain Engineering", email = "engineering@ml-chain.network" },
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "pynacl",
    "optuna",
    "requests",
    "unittest-xml-reporting",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
